Making a component the true component.
