+++ components/interop/interop-service/src/main/java/net/corda/interop/service/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * This package is only exported because an otherwise internal interface is being accessed from an example application outside this module.
+ * An `impl` package has been added underneath this package to keep OSGI happy.
+ */
+@Export
+package net.corda.interop.service;
+
+import org.osgi.annotation.bundle.Export;
